code cleanup, split all agg classes into individual files
