+++ src/seaq/Aggregations/IAggregationField.cs
@@ -1,0 +1,7 @@
+﻿namespace seaq
+{
+    public interface IAggregationField
+    {
+        public string FieldName { get; }
+    }
+}
