+++ src/seaq/Aggregations/IAggregationResult.cs
@@ -1,0 +1,7 @@
+﻿namespace seaq
+{
+    public interface IAggregationResult
+    {
+
+    }
+}
